+++ Changes.md
@@ -4,6 +4,7 @@
 Improvements
 ------------
 
+- LightToCamera : Added `distantAperture` and `clippingPlanes` plugs.
 - Arnold Render : Changed warnings for invalid mesh lights to be one descriptive warning per light, instead of repeating an unclear warning for every surface that links to the light.
 
 Fixes